# internal/auth/entities/userEntity.go
package entities

import "time"

type User struct {
	ID        string    `gorm:"primaryKey" json:"id"`
	Email     string    `json:"email"`
	Username  string    `json:"username"`
	Password  string    `json:"password"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}